feat(TAP-1190):Improve governance API (#36)

+++ client/src/rest/querys.rs
@@ -4,7 +4,7 @@ use utoipa::IntoParams;
 #[into_params(parameter_in = Query)]
 pub struct GetAllSubjectsQuery {
     pub from: Option<String>,
-    pub quantity: Option<usize>,
+    pub quantity: Option<i64>,
 }
 
 #[derive(Debug, Clone, Deserialize, IntoParams)]
